requirements.txt
fastapi[standard]==0.116.1
structlog==24.1.0
python-multipart==0.0.20
python-dotenv==1.1.1
Pillow==11.3.0
numpy==2.2.6
onnxruntime==1.22.1
tiktoken==0.11.0
aiokafka==0.12.0
huggingface-hub==0.34.4
uvicorn==0.35.0
gunicorn==22.0.0
httptools==0.6.4
websockets==15.0.1
watchfiles==1.1.0
pytest==8.4.2
flake8==7.3.0
autopep8==2.3.2
httpx==0.28.1
pydantic-settings==2.2.1
cachetools==5.3.3
redis==5.0.4
fastapi-limiter==0.1.6
prometheus-client==0.20.0
